Combine test_mini_config and test_enkf_main
This was really a test of enkf_main and should therefore
be in test_enkf_main
